% Generated by roxygen2 (4.0.0): do not edit by hand
\name{berlin_data}
\alias{berlin_data}
\title{Queries daten.berlin.de}
\usage{
berlin_data(query = NA, url = NA)
}
\arguments{
  \item{query}{a query string to search daten.berlin.de}

  \item{url}{a url pointing to a conrete dataset}
}
\description{
Only one of the parameters can have a value.
}

