% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{cached_datasets_feed}
\alias{cached_datasets_feed}
\title{Cached version of daten.berlin.de dataset feed.
Stored as XML tree in text form}
\format{XML tree with timestamp attribute 'last_updated'}
\source{
\url{http://daten.berlin.de/datensaetze/rss.xml}
}
\description{
Cached version of daten.berlin.de dataset feed.
Stored as XML tree in text form
}
\keyword{datasets}

