% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.berlin_data_resource}
\alias{download.berlin_data_resource}
\title{Downloads a resource}
\usage{
\method{download}{berlin_data_resource}(x, message.on.fail = TRUE,
  message.on.succeed = TRUE, ...)
}
\arguments{
\item{x}{a resource of type berlin_data_resource}

\item{message.on.fail}{logical: show message on download failure?}

\item{message.on.succeed}{logical: show message on sucessful download?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads a resource
}

