% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.CSV}
\alias{download.CSV}
\title{Downloads CSV}
\usage{
\method{download}{CSV}(resource, ..., sep = ";")
}
\arguments{
\item{resource}{url of csv file location}

\item{sep}{field separator}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads CSV
}

