% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.default}
\alias{download.default}
\title{Default fallback method for download
Indicates unsupported file format}
\usage{
\method{download}{default}(x, message.on.fail = TRUE, ...)
}
\arguments{
\item{x}{a resource which can be downloaded}

\item{...}{optional additional arguments}

\item{message.on.fail}{logical: show message indicating failure to download?}
}
\description{
Default fallback method for download
Indicates unsupported file format
}

