% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.JSON}
\alias{download.JSON}
\title{Downloads json file}
\usage{
\method{download}{JSON}(resource, ..., parse.to.df = TRUE)
}
\arguments{
\item{resource}{url of json file location}

\item{parse.to.df}{logical: should the function try to parse the JSON output into a data.frame?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads json file
}

