% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.TXT}
\alias{download.TXT}
\title{Downloads TXT}
\usage{
\method{download}{TXT}(x, message.on.fail = TRUE, ..., sep = ",")
}
\arguments{
\item{x}{url of txt file location}

\item{sep}{field separator}

\item{message.on.fail}{logical: show message on download failure?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads TXT
}

