% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.XLS}
\alias{download.XLS}
\title{Downloads xls file
Interestingly, these appear to actually be HTML files.}
\usage{
\method{download}{XLS}(resource, ..., parse.to.df = TRUE)
}
\arguments{
\item{resource}{url of xls file location}

\item{parse.to.df}{logical: should the function try to parse the html output into a data.frame?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads xls file
Interestingly, these appear to actually be HTML files.
}

