% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.XML}
\alias{download.XML}
\title{Downloads xml file}
\usage{
\method{download}{XML}(x, message.on.fail = TRUE, ..., parse.to.df = TRUE)
}
\arguments{
\item{x}{url of xml file location}

\item{message.on.fail}{logical: show message on download failure?}

\item{parse.to.df}{logical: should the function try to parse the XML output into a data.frame?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads xml file
}

