% Generated by roxygen2 (4.0.0): do not edit by hand
\name{downloadFormat}
\alias{downloadFormat}
\title{Calls appropriate function for download format}
\usage{
downloadFormat(url, format, ...)
}
\arguments{
\item{url}{url of file location}

\item{format}{string specifying download format, in c("CSV", "JSON", "XML")}

\item{...}{optional additional arguments to download function}
}
\description{
Calls appropriate function for download format
}

