% Generated by roxygen2 (4.0.0): do not edit by hand
\name{downloadJSON}
\alias{downloadJSON}
\title{Downloads json file}
\usage{
downloadJSON(json.url, parse.to.df = TRUE, ...)
}
\arguments{
\item{json.url}{url of json file location}

\item{parse.to.df}{logical: should the function try to parse the JSON output into a data.frame?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads json file
}

