% Generated by roxygen2 (4.0.0): do not edit by hand
\name{downloadXML}
\alias{downloadXML}
\title{Downloads xml file}
\usage{
downloadXML(xml.url, parse.to.df = TRUE, ...)
}
\arguments{
\item{xml.url}{url of xml file location}

\item{parse.to.df}{logical: should the function try to parse the XML output into a data.frame?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads xml file
}

