% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{feed}
\alias{feed}
\title{Cached version of the RSS feed}
\format{An XML document}
\description{
Cached version of the RSS feed
}
\keyword{Gedenken,}
\keyword{Geschichte,}
\keyword{Judenverfolgung}
\keyword{Nationalsozialismus,}
\keyword{Stolpersteine,}

