% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getDatasetMetaData}
\alias{getDatasetMetaData}
\title{Gets metadata for a dataset or list of datasets}
\usage{
getDatasetMetaData(where)
}
\arguments{
\item{where}{where to look for metadata: can be a URL, \code{berlin_data_dataset_info}, or \code{berlin_data_list}}
}
\description{
Gets metadata for a dataset or list of datasets
}

