% Generated by roxygen2 (4.0.0): do not edit by hand
\name{load_metadata}
\alias{load_metadata}
\title{Parses and downloads the meta data for a dataset}
\usage{
load_metadata(dataset_url)
}
\arguments{
  \item{dataset_url}{the url of the dataset}
}
\description{
Parses and downloads the meta data for a dataset
}

