% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parseMetaData}
\alias{parseMetaData}
\title{Parses and downloads the meta data for a dataset}
\usage{
parseMetaData(dataset_url)
}
\arguments{
\item{dataset_url}{the url of the dataset}
}
\description{
Parses and downloads the meta data for a dataset
}

