% Generated by roxygen2 (4.0.0): do not edit by hand
\name{resources.berlin_data_dataset}
\alias{resources.berlin_data_dataset}
\title{Gets the resources from a dataset}
\usage{
\method{resources}{berlin_data_dataset}(object, ...)
}
\arguments{
\item{object}{a dataset with a list of resources}

\item{...}{optional additional arguments}
}
\description{
Gets the resources from a dataset
}

