% Generated by roxygen2 (4.0.0): do not edit by hand
\name{resources}
\alias{resources}
\title{Gets the resources from an object}
\usage{
resources(object, ...)
}
\arguments{
\item{object}{an object with resources}

\item{...}{optional additional arguments to methods}
}
\description{
Gets the resources from an object
}

