% Generated by roxygen2 (4.0.0): do not edit by hand
\name{searchBerlinData}
\alias{searchBerlinData}
\title{Queries daten.berlin.de}
\usage{
searchBerlinData(query = NA)
}
\arguments{
\item{query}{a query string to search daten.berlin.de}
}
\description{
Only one of the parameters can have a value.
}
\examples{
result <- searchBerlinData(query = "stolpersteine")
dataset <- parseMetaData(result[[2]]$link)
resource_list <- resources(dataset)
data <- download(resource_list[[1]])
}

