% Generated by roxygen2 (4.0.0): do not edit by hand
\name{searchBerlinDatasets}
\alias{searchBerlinDatasets}
\title{Queries daten.berlin.de}
\usage{
searchBerlinDatasets(query, ...)
}
\arguments{
\item{query}{a query string to search daten.berlin.de}

\item{...}{optional additional arguments}
}
\description{
Queries daten.berlin.de
}
\examples{
result <- searchBerlinDatasets(query = "stolpersteine")
summary(result)
dataset <- getDatasetMetaData(result[[2]])
summary(dataset)
resource_list <- resources(dataset)
summary(resource_list)
data <- download(resource_list[[1]], sep=',')
}

