% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{stolpersteine}
\alias{stolpersteine}
\title{Information on the Stopersteine in Berlin.}
\format{A data frame with 5328 rows and 14 variables}
\source{
Website of the project: \url{http://www.stolpersteine-berlin.de/}

Website of data:
\url{http://daten.berlin.de/datensaetze/liste-der-stolpersteine-berlin}
Version: 2013-08-16 (6f2faa4e-9249-4e1b-a45b-b30799e20768)
}
\usage{
data(stolpersteine)
}
\description{
A dataset containing information on all persons on the famous Stolpersteine
in Berlin.
}
\details{
LICENSE:
Attribution 3.0 Unported (CC BY 3.0)
\url{http://creativecommons.org/licenses/by/3.0/}

Fields:
\itemize{
\item Firstname. the first name of the person. as factor.
\item Familyname. the family name of the person. as factor.
\item Lastname. lastname of the person. as factor.
\item Birthday. date of birth. as date. contains NAs.
\item Address. the address of the person. as factor.
\item District. the destrict of Berlin where the person lived as factor.
\item DayOfDeportation. date of first deportation. contains NAs.
\item DestinationOfDeportation. destination of deportation. as factor. contains NAs.
\item DayOfDeportation2. date of second deportation. contains NAs.
\item DestinationOfDeportation2.  destination of deportation. as factor. contains NAs.
\item DayOfDeportation3. date of third deportation. contains NAs.
\item DestinationOfDeportation3.  destination of deportation. as factor. contains NAs.
\item DayOfDeath. date of death. contains NAs.
\item PlaceOfDeath. place of death as factor.
}
}
\examples{
\dontrun{
# this is how the data was produced
stolpersteine <- read.csv("stolpersteine.csv", sep = ",", na.strings = c(""),
                          fileEncoding = "utf8")
stolpersteine$Geburtstag <- strptime(stolpersteine$Geburtstag, "\%d/\%m/\%Y")
stolpersteine$Todestag <- strptime(stolpersteine$Todestag, "\%d/\%m/\%Y")
stolpersteine$Deportationstag <- strptime(stolpersteine$Deportationstag, "\%d/\%m/\%Y")
stolpersteine$X2.Deportationstag <- strptime(stolpersteine$X2.Deportationstag, "\%d/\%m/\%Y")
stolpersteine$X3.Deportationstag <- strptime(stolpersteine$X3.Deportationstag, "\%d/\%m/\%Y")
colnames(stolpersteine) <- c("Firstname", "Familiyname", "Lastname",
                             "Birthday", "Address", "District",
                             "DayOfDeportation", "DestinationOfDeportation",
                             "DayOfDeportation2", "DestinationOfDeportation2",
                             "DayOfDeportation3", "DestinationOfDeportation3",
                             "DayOfDeath", "PlaceOfDeath")
save(stolpersteine, file = "data/stolpersteine.rda")
}
}
\keyword{Gedenken,}
\keyword{Geschichte,}
\keyword{Judenverfolgung}
\keyword{Nationalsozialismus,}
\keyword{Stolpersteine,}

