% Generated by roxygen2 (4.0.0): do not edit by hand
\name{updateCachedFeed}
\alias{updateCachedFeed}
\title{Updates cached datasets_feed, in data/cached_datasets_feed.rda, from daten.berlin.de}
\usage{
updateCachedFeed(feed.url = "http://daten.berlin.de/datensaetze/rss.xml")
}
\description{
Updates cached datasets_feed, in data/cached_datasets_feed.rda, from daten.berlin.de
}

