% Generated by roxygen2 (4.0.0): do not edit by hand
\name{berlin_data}
\alias{berlin_data}
\title{Queries daten.berlin.de}
\usage{
berlin_data(query = NA, url = NA)
}
\arguments{
  \item{query}{a query string to search daten.berlin.de}

  \item{url}{a url pointing to a concrete dataset}
}
\description{
Only one of the parameters can have a value.
}
\examples{
result <- berlin_data(query = "stolpersteine")
dataset <- load_metadata(result[[2]]$link)
resources <- resources(dataset)
data <- download(resources[[1]])
}

