% Generated by roxygen2 (4.0.0): do not edit by hand
\name{download.default}
\alias{download.default}
\title{Default fallback method for download
Indicates unsupported file format}
\usage{
\method{download}{default}(resource, ...)
}
\arguments{
\item{resource}{a resource which can be downloaded}

\item{...}{optional additional arguments}
}
\description{
Default fallback method for download
Indicates unsupported file format
}

