% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download.JSON}
\alias{download.JSON}
\title{Downloads json file}
\usage{
\method{download}{JSON}(x, message.on.fail = TRUE, ..., parse.to.df = TRUE)
}
\arguments{
\item{x}{url of json file location}

\item{message.on.fail}{logical: show message on download failure?}

\item{parse.to.df}{logical: should the function try to parse the JSON output into a data.frame?}

\item{...}{optional additional arguments to download function}
}
\description{
Downloads json file
}

