% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDatasetMetaData}
\alias{getDatasetMetaData}
\title{Gets metadata for a dataset or list of datasets}
\usage{
getDatasetMetaData(where, ...)
}
\arguments{
\item{where}{where to look for metadata}

\item{...}{optional additional arguments to methods}
}
\description{
Gets metadata for a dataset or list of datasets
}
\examples{
# query
\dontrun{
result <- searchBerlinDatasets(query = "vornamen")
summary(result)

# options to get metadata:
# pass in URL
metadata <- getDatasetMetaData(result[[2]]$link)
# returns 'berlin_data_dataset' object with list of available resources
class(metadata)

# pass in object
metadata <- getDatasetMetaData(result[[2]])
# same result as passing in link

# pass in berlin_data_list with query results to get metadata
# for all objects in list
metadata <- getDatasetMetaData(result)
summary(metadata)
summary(metadata[[1]])
}
}

