% Generated by roxygen2 (4.0.0): do not edit by hand
\name{searchBerlinDatasets}
\alias{searchBerlinDatasets}
\title{Queries daten.berlin.de}
\usage{
searchBerlinDatasets(query)
}
\arguments{
\item{query}{a query string to search daten.berlin.de}
}
\description{
Only one of the parameters can have a value.
}
\examples{
result <- searchBerlinDatasets(query = "stolpersteine")
dataset <- getDatasetMetaData(result[[2]])
resource_list <- resources(dataset)
data <- download(resource_list[[1]])
}

