% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateCachedFeed}
\alias{updateCachedFeed}
\title{Updates cached_datasets_feed}
\usage{
updateCachedFeed(xml.url = "http://daten.berlin.de/datensaetze/rss.xml")
}
\arguments{
\item{xml.url}{the url to the rss feed}
}
\description{
Updates cached_datasets_feed
}

